refactor: delegate inventoryForProductConfigurations to simple-inventory plugin
